AC- refactoring and reformatting
