+++ openmrs-client/src/main/java/org/openmrs/mobile/activities/providerdashboard/ProviderDashboardPresenter.java
@@ -6,7 +6,7 @@ import android.content.Intent;
 import androidx.annotation.NonNull;
 
 import org.openmrs.mobile.activities.BasePresenter;
-import org.openmrs.mobile.listeners.retrofit.CustomApiCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.CustomApiCallback;
 import org.openmrs.mobile.api.RestApi;
 import org.openmrs.mobile.api.RestServiceBuilder;
 import org.openmrs.mobile.api.repository.ProviderRepository;
