+++ openmrs-client/src/main/java/org/openmrs/mobile/api/EncounterService.java
@@ -22,8 +22,8 @@ import org.openmrs.mobile.application.OpenMRS;
 import org.openmrs.mobile.dao.PatientDAO;
 import org.openmrs.mobile.dao.VisitDAO;
 import org.openmrs.mobile.databases.AppDatabase;
-import org.openmrs.mobile.listeners.retrofit.DefaultResponseCallback;
-import org.openmrs.mobile.listeners.retrofit.StartVisitResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.DefaultResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.StartVisitResponseCallback;
 import org.openmrs.mobile.models.Encounter;
 import org.openmrs.mobile.models.EncounterType;
 import org.openmrs.mobile.models.Encountercreate;
