+++ openmrs-client/src/main/java/org/openmrs/mobile/api/repository/PatientRepository.java
@@ -40,10 +40,10 @@ import org.openmrs.mobile.dao.EncounterCreateRoomDAO;
 import org.openmrs.mobile.dao.PatientDAO;
 import org.openmrs.mobile.databases.AppDatabase;
 import org.openmrs.mobile.databases.entities.LocationEntity;
-import org.openmrs.mobile.listeners.retrofit.DefaultResponseCallback;
-import org.openmrs.mobile.listeners.retrofit.DownloadPatientCallback;
-import org.openmrs.mobile.listeners.retrofit.PatientDeferredResponseCallback;
-import org.openmrs.mobile.listeners.retrofit.PatientResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.DefaultResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.DownloadPatientCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.PatientDeferredResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.PatientResponseCallback;
 import org.openmrs.mobile.models.Encountercreate;
 import org.openmrs.mobile.models.IdGenPatientIdentifiers;
 import org.openmrs.mobile.models.IdentifierType;
