+++ openmrs-client/src/main/java/org/openmrs/mobile/api/repository/ProviderRepository.java
@@ -26,7 +26,7 @@ import androidx.work.WorkManager;
 
 import org.jetbrains.annotations.NotNull;
 import org.openmrs.mobile.R;
-import org.openmrs.mobile.listeners.retrofit.CustomApiCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.CustomApiCallback;
 import org.openmrs.mobile.api.RestApi;
 import org.openmrs.mobile.api.RestServiceBuilder;
 import org.openmrs.mobile.api.workers.provider.AddProviderWorker;
@@ -36,8 +36,8 @@ import org.openmrs.mobile.application.OpenMRS;
 import org.openmrs.mobile.dao.ProviderRoomDAO;
 import org.openmrs.mobile.databases.AppDatabase;
 import org.openmrs.mobile.databases.entities.LocationEntity;
-import org.openmrs.mobile.listeners.retrofit.EncounterResponseCallback;
-import org.openmrs.mobile.listeners.retrofit.LocationResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.EncounterResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.LocationResponseCallback;
 import org.openmrs.mobile.models.Provider;
 import org.openmrs.mobile.models.Resource;
 import org.openmrs.mobile.models.Results;
