+++ openmrs-client/src/main/java/org/openmrs/mobile/api/repository/VisitRepository.java
@@ -24,10 +24,10 @@ import org.openmrs.mobile.dao.EncounterDAO;
 import org.openmrs.mobile.dao.LocationDAO;
 import org.openmrs.mobile.dao.VisitDAO;
 import org.openmrs.mobile.databases.AppDatabase;
-import org.openmrs.mobile.listeners.retrofit.DefaultResponseCallback;
-import org.openmrs.mobile.listeners.retrofit.VisitsResponseCallback;
-import org.openmrs.mobile.listeners.retrofit.GetVisitTypeCallback;
-import org.openmrs.mobile.listeners.retrofit.StartVisitResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.DefaultResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.VisitsResponseCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.GetVisitTypeCallback;
+import org.openmrs.mobile.listeners.retrofitcallbacks.StartVisitResponseCallback;
 import org.openmrs.mobile.models.Encounter;
 import org.openmrs.mobile.models.Patient;
 import org.openmrs.mobile.models.Results;
