+++ openmrs-client/src/main/java/org/openmrs/mobile/listeners/retrofitcallbacks/DefaultResponseCallback.java
@@ -12,22 +12,20 @@
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 
-package org.openmrs.mobile.listeners.retrofit;
+package org.openmrs.mobile.listeners.retrofitcallbacks;
 
 /**
  * this method {@link #onResponse()} () onResponse}
  * will be overridden in custom callback methods if the response matches with the use case.
  * otherwise  method will be overloaded in the custom callbacks
- *
  * this method {@link #onErrorResponse(String)}() onErrorResponse}
  * will be overridden in custom callback methods since it has uniform usage throughout the
  * app in various repositories
- *
  */
 public interface DefaultResponseCallback {
-
     default void onResponse() {
     }
 
-    default void onErrorResponse(String errorMessage){};
+    default void onErrorResponse(String errorMessage) {
+    }
 }
