+++ openmrs-client/src/main/java/org/openmrs/mobile/listeners/retrofitcallbacks/DownloadPatientCallback.java
@@ -12,18 +12,12 @@
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 
-package org.openmrs.mobile.listeners.retrofit;
+package org.openmrs.mobile.listeners.retrofitcallbacks;
 
-import org.openmrs.mobile.api.promise.SimpleDeferredObject;
 import org.openmrs.mobile.models.Patient;
 
-public interface PatientDeferredResponseCallback extends DefaultResponseCallback {
-    default void onResponse(SimpleDeferredObject<Patient> response) {
-    }
+public interface DownloadPatientCallback extends DefaultResponseCallback {
+    void onPatientDownloaded(Patient patient);
 
-    default void onErrorResponse(String errorMessage, SimpleDeferredObject<Patient> errorResponse) {
-    }
-
-    default void onNotifyResponse(String notifyMessage) {
-    }
+    void onPatientPhotoDownloaded(Patient patient);
 }
