+++ openmrs-client/src/main/java/org/openmrs/mobile/listeners/retrofitcallbacks/PatientDeferredResponseCallback.java
@@ -12,12 +12,18 @@
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 
-package org.openmrs.mobile.listeners.retrofit;
+package org.openmrs.mobile.listeners.retrofitcallbacks;
 
+import org.openmrs.mobile.api.promise.SimpleDeferredObject;
 import org.openmrs.mobile.models.Patient;
 
-public interface DownloadPatientCallback extends DefaultResponseCallback {
-    void onPatientDownloaded(Patient patient);
+public interface PatientDeferredResponseCallback extends DefaultResponseCallback {
+    default void onResponse(SimpleDeferredObject<Patient> response) {
+    }
 
-    void onPatientPhotoDownloaded(Patient patient);
+    default void onErrorResponse(String errorMessage, SimpleDeferredObject<Patient> errorResponse) {
+    }
+
+    default void onNotifyResponse(String notifyMessage) {
+    }
 }
