+++ openmrs-client/src/main/java/org/openmrs/mobile/listeners/retrofitcallbacks/StartVisitResponseCallback.java
@@ -12,11 +12,8 @@
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 
-package org.openmrs.mobile.listeners.retrofit
+package org.openmrs.mobile.listeners.retrofitcallbacks;
 
-import org.openmrs.mobile.databases.entities.LocationEntity
-
-@JvmSuppressWildcards
-interface LocationResponseCallback : DefaultResponseCallback {
-    fun onResponse(locationList: List<LocationEntity?>?)
+public interface StartVisitResponseCallback extends DefaultResponseCallback {
+    void onStartVisitResponse(long id);
 }
