+++ openmrs-client/src/main/java/org/openmrs/mobile/listeners/retrofitcallbacks/VisitsResponseCallback.java
@@ -12,10 +12,10 @@
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
 
-package org.openmrs.mobile.listeners.retrofit;
+package org.openmrs.mobile.listeners.retrofitcallbacks;
 
-import org.openmrs.mobile.models.VisitType;
+public interface VisitsResponseCallback {
+    void onSuccess(String response);
 
-public interface GetVisitTypeCallback extends DefaultResponseCallback {
-    void onGetVisitTypeResponse(VisitType visitType);
+    void onFailure(String errorMessage);
 }
